Added a representation of the results:
Main:
Added a function for drawing the tracks that match the user's preferences, with colors and popups containing details about the tracks.

OsmTrack, OsmDbGenerator:
Changed the osmDb file structure to be a dictionary of the form: {tracks {id: {}}} (we can now get the data of a track in O(1) by using it's id) in order to have fast access to the data we want to put in the track's popup.

ResultsTests:
Modified pretty_print_results to be consistent with the db structure change.